+++ worlds/banjo_tooie/tests/__init__.py
@@ -1,4 +1,0 @@
-from test.bases import WorldTestBase
-
-class BanjoTooieTestBase(WorldTestBase):
-    game = "Banjo-Tooie"